Move isValidHost to ipUtil.js
